+++ src/Fp/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Fix xor arm intrinsic
+- Initialize Processor.LittleEndian to true in constructor
 
 ## [0.23.2] 2022-1-11
 ### Changed